test: app state
