+++ lib/application/redux/actions/update_user_profile_action.dart
@@ -2,6 +2,7 @@
 import 'package:async_redux/async_redux.dart';
 import 'package:domain_objects/entities.dart';
 import 'package:healthcloud/application/redux/states/app_state.dart';
+import 'package:healthcloud/application/redux/states/staff_state.dart';
 import 'package:healthcloud/application/redux/states/user_state.dart';
 
 /// - [userProfile]
