+++ lib/application/redux/states/auth_state.dart
@@ -1,0 +1,17 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+import 'package:healthcloud/domain/core/entities/core/auth_credentials.dart';
+
+part 'auth_state.freezed.dart';
+part 'auth_state.g.dart';
+
+@freezed
+class AuthState with _$AuthState {
+  factory AuthState({AuthCredentials? credentials}) = _AuthState;
+
+  factory AuthState.fromJson(Map<String, dynamic> json) =>
+      _$AuthStateFromJson(json);
+
+  factory AuthState.initial() => AuthState(
+        credentials: AuthCredentials.initial(),
+      );
+}
