+++ lib/application/redux/states/auth_state.freezed.dart
@@ -1,0 +1,174 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+
+part of 'auth_state.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+AuthState _$AuthStateFromJson(Map<String, dynamic> json) {
+  return _AuthState.fromJson(json);
+}
+
+/// @nodoc
+class _$AuthStateTearOff {
+  const _$AuthStateTearOff();
+
+  _AuthState call({AuthCredentials? credentials}) {
+    return _AuthState(
+      credentials: credentials,
+    );
+  }
+
+  AuthState fromJson(Map<String, Object> json) {
+    return AuthState.fromJson(json);
+  }
+}
+
+/// @nodoc
+const $AuthState = _$AuthStateTearOff();
+
+/// @nodoc
+mixin _$AuthState {
+  AuthCredentials? get credentials => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $AuthStateCopyWith<AuthState> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $AuthStateCopyWith<$Res> {
+  factory $AuthStateCopyWith(AuthState value, $Res Function(AuthState) then) =
+      _$AuthStateCopyWithImpl<$Res>;
+  $Res call({AuthCredentials? credentials});
+
+  $AuthCredentialsCopyWith<$Res>? get credentials;
+}
+
+/// @nodoc
+class _$AuthStateCopyWithImpl<$Res> implements $AuthStateCopyWith<$Res> {
+  _$AuthStateCopyWithImpl(this._value, this._then);
+
+  final AuthState _value;
+  // ignore: unused_field
+  final $Res Function(AuthState) _then;
+
+  @override
+  $Res call({
+    Object? credentials = freezed,
+  }) {
+    return _then(_value.copyWith(
+      credentials: credentials == freezed
+          ? _value.credentials
+          : credentials // ignore: cast_nullable_to_non_nullable
+              as AuthCredentials?,
+    ));
+  }
+
+  @override
+  $AuthCredentialsCopyWith<$Res>? get credentials {
+    if (_value.credentials == null) {
+      return null;
+    }
+
+    return $AuthCredentialsCopyWith<$Res>(_value.credentials!, (value) {
+      return _then(_value.copyWith(credentials: value));
+    });
+  }
+}
+
+/// @nodoc
+abstract class _$AuthStateCopyWith<$Res> implements $AuthStateCopyWith<$Res> {
+  factory _$AuthStateCopyWith(
+          _AuthState value, $Res Function(_AuthState) then) =
+      __$AuthStateCopyWithImpl<$Res>;
+  @override
+  $Res call({AuthCredentials? credentials});
+
+  @override
+  $AuthCredentialsCopyWith<$Res>? get credentials;
+}
+
+/// @nodoc
+class __$AuthStateCopyWithImpl<$Res> extends _$AuthStateCopyWithImpl<$Res>
+    implements _$AuthStateCopyWith<$Res> {
+  __$AuthStateCopyWithImpl(_AuthState _value, $Res Function(_AuthState) _then)
+      : super(_value, (v) => _then(v as _AuthState));
+
+  @override
+  _AuthState get _value => super._value as _AuthState;
+
+  @override
+  $Res call({
+    Object? credentials = freezed,
+  }) {
+    return _then(_AuthState(
+      credentials: credentials == freezed
+          ? _value.credentials
+          : credentials // ignore: cast_nullable_to_non_nullable
+              as AuthCredentials?,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_AuthState implements _AuthState {
+  _$_AuthState({this.credentials});
+
+  factory _$_AuthState.fromJson(Map<String, dynamic> json) =>
+      _$_$_AuthStateFromJson(json);
+
+  @override
+  final AuthCredentials? credentials;
+
+  @override
+  String toString() {
+    return 'AuthState(credentials: $credentials)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other is _AuthState &&
+            (identical(other.credentials, credentials) ||
+                const DeepCollectionEquality()
+                    .equals(other.credentials, credentials)));
+  }
+
+  @override
+  int get hashCode =>
+      runtimeType.hashCode ^ const DeepCollectionEquality().hash(credentials);
+
+  @JsonKey(ignore: true)
+  @override
+  _$AuthStateCopyWith<_AuthState> get copyWith =>
+      __$AuthStateCopyWithImpl<_AuthState>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$_$_AuthStateToJson(this);
+  }
+}
+
+abstract class _AuthState implements AuthState {
+  factory _AuthState({AuthCredentials? credentials}) = _$_AuthState;
+
+  factory _AuthState.fromJson(Map<String, dynamic> json) =
+      _$_AuthState.fromJson;
+
+  @override
+  AuthCredentials? get credentials => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$AuthStateCopyWith<_AuthState> get copyWith =>
+      throw _privateConstructorUsedError;
+}
