+++ lib/application/redux/states/home_state.dart
@@ -1,0 +1,18 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'home_state.freezed.dart';
+part 'home_state.g.dart';
+
+@freezed
+class HomeState with _$HomeState {
+  factory HomeState() = _HomeState;
+
+  factory HomeState.fromJson(Map<String, dynamic> json) =>
+      _$HomeStateFromJson(json);
+
+  factory HomeState.initial() => HomeState(
+
+      ///   - SurveyState
+      /// ServiceRequests
+      );
+}
