+++ lib/application/redux/states/onboarding_state.dart
@@ -1,0 +1,18 @@
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'onboarding_state.freezed.dart';
+part 'onboarding_state.g.dart';
+
+@freezed
+class OnboardingState with _$OnboardingState {
+  factory OnboardingState() = _OnboardingState;
+
+  factory OnboardingState.fromJson(Map<String, dynamic> json) =>
+      _$OnboardingStateFromJson(json);
+
+  factory OnboardingState.initial() => OnboardingState(
+
+      ///   - Login state
+      ///   - create pin
+      );
+}
