+++ lib/application/redux/states/onboarding_state.freezed.dart
@@ -1,0 +1,108 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+
+part of 'onboarding_state.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+OnboardingState _$OnboardingStateFromJson(Map<String, dynamic> json) {
+  return _OnboardingState.fromJson(json);
+}
+
+/// @nodoc
+class _$OnboardingStateTearOff {
+  const _$OnboardingStateTearOff();
+
+  _OnboardingState call() {
+    return _OnboardingState();
+  }
+
+  OnboardingState fromJson(Map<String, Object> json) {
+    return OnboardingState.fromJson(json);
+  }
+}
+
+/// @nodoc
+const $OnboardingState = _$OnboardingStateTearOff();
+
+/// @nodoc
+mixin _$OnboardingState {
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $OnboardingStateCopyWith<$Res> {
+  factory $OnboardingStateCopyWith(
+          OnboardingState value, $Res Function(OnboardingState) then) =
+      _$OnboardingStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$OnboardingStateCopyWithImpl<$Res>
+    implements $OnboardingStateCopyWith<$Res> {
+  _$OnboardingStateCopyWithImpl(this._value, this._then);
+
+  final OnboardingState _value;
+  // ignore: unused_field
+  final $Res Function(OnboardingState) _then;
+}
+
+/// @nodoc
+abstract class _$OnboardingStateCopyWith<$Res> {
+  factory _$OnboardingStateCopyWith(
+          _OnboardingState value, $Res Function(_OnboardingState) then) =
+      __$OnboardingStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$OnboardingStateCopyWithImpl<$Res>
+    extends _$OnboardingStateCopyWithImpl<$Res>
+    implements _$OnboardingStateCopyWith<$Res> {
+  __$OnboardingStateCopyWithImpl(
+      _OnboardingState _value, $Res Function(_OnboardingState) _then)
+      : super(_value, (v) => _then(v as _OnboardingState));
+
+  @override
+  _OnboardingState get _value => super._value as _OnboardingState;
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_OnboardingState implements _OnboardingState {
+  _$_OnboardingState();
+
+  factory _$_OnboardingState.fromJson(Map<String, dynamic> json) =>
+      _$_$_OnboardingStateFromJson(json);
+
+  @override
+  String toString() {
+    return 'OnboardingState()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _OnboardingState);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$_$_OnboardingStateToJson(this);
+  }
+}
+
+abstract class _OnboardingState implements OnboardingState {
+  factory _OnboardingState() = _$_OnboardingState;
+
+  factory _OnboardingState.fromJson(Map<String, dynamic> json) =
+      _$_OnboardingState.fromJson;
+}
