+++ lib/application/redux/states/survey_state.freezed.dart
@@ -1,0 +1,106 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
+
+part of 'survey_state.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+SurveyState _$SurveyStateFromJson(Map<String, dynamic> json) {
+  return _SurveyState.fromJson(json);
+}
+
+/// @nodoc
+class _$SurveyStateTearOff {
+  const _$SurveyStateTearOff();
+
+  _SurveyState call() {
+    return _SurveyState();
+  }
+
+  SurveyState fromJson(Map<String, Object> json) {
+    return SurveyState.fromJson(json);
+  }
+}
+
+/// @nodoc
+const $SurveyState = _$SurveyStateTearOff();
+
+/// @nodoc
+mixin _$SurveyState {
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $SurveyStateCopyWith<$Res> {
+  factory $SurveyStateCopyWith(
+          SurveyState value, $Res Function(SurveyState) then) =
+      _$SurveyStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class _$SurveyStateCopyWithImpl<$Res> implements $SurveyStateCopyWith<$Res> {
+  _$SurveyStateCopyWithImpl(this._value, this._then);
+
+  final SurveyState _value;
+  // ignore: unused_field
+  final $Res Function(SurveyState) _then;
+}
+
+/// @nodoc
+abstract class _$SurveyStateCopyWith<$Res> {
+  factory _$SurveyStateCopyWith(
+          _SurveyState value, $Res Function(_SurveyState) then) =
+      __$SurveyStateCopyWithImpl<$Res>;
+}
+
+/// @nodoc
+class __$SurveyStateCopyWithImpl<$Res> extends _$SurveyStateCopyWithImpl<$Res>
+    implements _$SurveyStateCopyWith<$Res> {
+  __$SurveyStateCopyWithImpl(
+      _SurveyState _value, $Res Function(_SurveyState) _then)
+      : super(_value, (v) => _then(v as _SurveyState));
+
+  @override
+  _SurveyState get _value => super._value as _SurveyState;
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_SurveyState implements _SurveyState {
+  _$_SurveyState();
+
+  factory _$_SurveyState.fromJson(Map<String, dynamic> json) =>
+      _$_$_SurveyStateFromJson(json);
+
+  @override
+  String toString() {
+    return 'SurveyState()';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) || (other is _SurveyState);
+  }
+
+  @override
+  int get hashCode => runtimeType.hashCode;
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$_$_SurveyStateToJson(this);
+  }
+}
+
+abstract class _SurveyState implements SurveyState {
+  factory _SurveyState() = _$_SurveyState;
+
+  factory _SurveyState.fromJson(Map<String, dynamic> json) =
+      _$_SurveyState.fromJson;
+}
