+++ lib/infrastructure/repository/database_state_persistor.dart
@@ -5,6 +5,12 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter_graphql_client/graph_sqlite.dart';
 import 'package:healthcloud/application/redux/states/app_state.dart';
 import 'package:healthcloud/application/redux/states/home/bottom_nav_state.dart';
+import 'package:healthcloud/application/redux/states/home_state.dart';
+import 'package:healthcloud/application/redux/states/misc_state.dart';
+import 'package:healthcloud/application/redux/states/onboarding_state.dart';
+import 'package:healthcloud/application/redux/states/staff_state.dart';
+import 'package:healthcloud/application/redux/states/survey_requests_state.dart';
+import 'package:healthcloud/application/redux/states/survey_state.dart';
 import 'package:healthcloud/infrastructure/repository/database_base.dart';
 import 'package:healthcloud/infrastructure/repository/database_mobile.dart';
 import 'package:healthcloud/infrastructure/repository/initialize_db.dart';
