+++ test/unit_tests/application/redux/states/app_state_test.dart
@@ -1,7 +1,9 @@
-// Package imports:
 import 'package:healthcloud/application/redux/states/app_state.dart';
+import 'package:healthcloud/application/redux/states/staff_state.dart';
 import 'package:healthcloud/application/redux/states/user_state.dart';
 import 'package:test/test.dart';
+
+import '../../../../mocks/mocks.dart';
 
 void main() {
   group('AppState', () {
@@ -16,6 +18,13 @@ void main() {
       );
 
       expect(state, isNot(equals(AppState.initial())));
+    });
+
+    test('fromJson', () {
+      expect(
+        AppState.fromJson(appstateMock),
+        AppState.initial().copyWith(wait: null),
+      );
     });
   });
 }
