+++ test/unit_tests/application/redux/states/misc_state_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:healthcloud/application/redux/states/misc_state.dart';
+
+import '../../../../mocks/mocks.dart';
+
+void main() {
+  group('MiscState', () {
+    test('fromJson', () {
+      expect(
+        MiscState.fromJson(mockMiscState),
+        MiscState.initial(),
+      );
+    });
+  });
+}
