+++ test/unit_tests/application/redux/states/onboarding_state_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:healthcloud/application/redux/states/onboarding_state.dart';
+
+void main() {
+  group('OnboardingState', () {
+    test('fromJson', () {
+      // OnboardingState is empty for now. Test should probably change once
+      // class is implemented
+      expect(
+        OnboardingState.fromJson(<String, dynamic>{}),
+        isA<OnboardingState>(),
+      );
+    });
+  });
+}
