+++ test/unit_tests/application/redux/states/staff_state_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:healthcloud/application/redux/states/staff_state.dart';
+
+import '../../../../mocks/mocks.dart';
+
+void main() {
+  group('StaffState', () {
+    test('fromJson', () {
+      expect(
+        StaffState.fromJson(mockStaffState),
+        StaffState.initial(),
+      );
+    });
+  });
+}
