+++ test/unit_tests/application/redux/states/survey_requests_state_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:healthcloud/application/redux/states/survey_requests_state.dart';
+
+void main() {
+  group('ServiceRequestState', () {
+    test('fromJson', () {
+      // ServiceRequestState is empty for now. Test should probably change once
+      // class is implemented
+      expect(
+        ServiceRequestState.fromJson(<String, dynamic>{}),
+        isA<ServiceRequestState>(),
+      );
+    });
+  });
+}
