+++ test/unit_tests/application/redux/states/survey_state_test.dart
@@ -1,0 +1,15 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:healthcloud/application/redux/states/survey_state.dart';
+
+void main() {
+  group('SurveyState', () {
+    test('fromJson', () {
+      // SurveyState is empty for now. Test should probably change once
+      // class is implemented
+      expect(
+        SurveyState.fromJson(<String, dynamic>{}),
+        isA<SurveyState>(),
+      );
+    });
+  });
+}
